feat: modify lambda types and interactions

+++ aws-lambda/pkg/types/action.go
@@ -1,6 +1,6 @@
 package types
 
 type Sum struct {
-	A int `json:"a"`
-	B int `json:"b"`
+	A int64 `json:"a"`
+	B int64 `json:"b"`
 }
